#  log: performance improvements in gelf field parsing