```
name                                      old time/op    new time/op    delta
GelfBackendParseFieldsNoFields-4            12.7ns ± 1%     9.1ns ± 1%  -28.10%  (p=0.002 n=6+6)
GelfBackendInvalidFields-4                   417ns ± 2%     218ns ± 1%  -47.72%  (p=0.000 n=6+5)
GelfBackendValidFields-4                    1.11µs ± 0%    0.60µs ± 1%  -46.04%  (p=0.002 n=6+6)
GelfBackendHugeValidFields-4                2.09µs ± 1%    0.65µs ± 0%  -69.10%  (p=0.016 n=5+4)
GelfBackendHugeValidFieldsBigWhitelist-4    8.47µs ± 2%    0.71µs ± 6%  -91.61%  (p=0.002 n=6+6)
GelfBackendHugeInvalid-4                    9.56µs ± 1%    0.59µs ± 1%  -93.80%  (p=0.002 n=6+6)

name                                      old alloc/op   new alloc/op   delta
GelfBackendParseFieldsNoFields-4             0.00B          0.00B          ~     (all equal)
GelfBackendInvalidFields-4                   0.00B          0.00B          ~     (all equal)
GelfBackendValidFields-4                      120B ± 0%       64B ± 0%  -46.67%  (p=0.002 n=6+6)
GelfBackendHugeValidFields-4                  144B ± 0%       80B ± 0%  -44.44%  (p=0.002 n=6+6)
GelfBackendHugeValidFieldsBigWhitelist-4      144B ± 0%       80B ± 0%  -44.44%  (p=0.002 n=6+6)
GelfBackendHugeInvalid-4                     0.00B          0.00B          ~     (all equal)

name                                      old allocs/op  new allocs/op  delta
GelfBackendParseFieldsNoFields-4              0.00           0.00          ~     (all equal)
GelfBackendInvalidFields-4                    0.00           0.00          ~     (all equal)
GelfBackendValidFields-4                      8.00 ± 0%      4.00 ± 0%  -50.00%  (p=0.002 n=6+6)
GelfBackendHugeValidFields-4                  10.0 ± 0%       5.0 ± 0%  -50.00%  (p=0.002 n=6+6)
GelfBackendHugeValidFieldsBigWhitelist-4      10.0 ± 0%       5.0 ± 0%  -50.00%  (p=0.002 n=6+6)
GelfBackendHugeInvalid-4                      0.00           0.00          ~     (all equal)
```